refactor: 텍스트 복사 함수 분리
